requirements.txt
requests
watchdog
mysql-connector-python